+++ relay-server/src/services/outcome.rs
@@ -171,8 +171,11 @@ pub enum Outcome {
     RateLimited(Option<ReasonCode>),
 
     /// The event/metric has been cardinality limited.
+    ///
+    /// Contains the [ID](relay_cardinality::CardinalityLimit::id)
+    /// of the most specific [CardinalityLimit](relay_cardinality::CardinalityLimit) that was exceeded.
     #[cfg(feature = "processing")]
-    CardinalityLimited,
+    CardinalityLimited(String),
 
     /// The event has been discarded because of invalid data.
     Invalid(DiscardReason),
@@ -192,7 +195,7 @@ impl Outcome {
             Outcome::Filtered(_) | Outcome::FilteredSampling(_) => OutcomeId::FILTERED,
             Outcome::RateLimited(_) => OutcomeId::RATE_LIMITED,
             #[cfg(feature = "processing")]
-            Outcome::CardinalityLimited => OutcomeId::CARDINALITY_LIMITED,
+            Outcome::CardinalityLimited(_) => OutcomeId::CARDINALITY_LIMITED,
             Outcome::Invalid(_) => OutcomeId::INVALID,
             Outcome::Abuse => OutcomeId::ABUSE,
             Outcome::ClientDiscard(_) => OutcomeId::CLIENT_DISCARD,
@@ -210,7 +213,7 @@ impl Outcome {
                 code_opt.as_ref().map(|code| Cow::Borrowed(code.as_str()))
             }
             #[cfg(feature = "processing")]
-            Outcome::CardinalityLimited => None,
+            Outcome::CardinalityLimited(id) => Some(Cow::Borrowed(id)),
             Outcome::ClientDiscard(ref discard_reason) => Some(Cow::Borrowed(discard_reason)),
             Outcome::Abuse => None,
             Outcome::Accepted => None,
@@ -242,7 +245,7 @@ impl fmt::Display for Outcome {
             Outcome::RateLimited(None) => write!(f, "rate limited"),
             Outcome::RateLimited(Some(reason)) => write!(f, "rate limited with reason {reason}"),
             #[cfg(feature = "processing")]
-            Outcome::CardinalityLimited => write!(f, "cardinality limited"),
+            Outcome::CardinalityLimited(id) => write!(f, "cardinality limited ({})", id),
             Outcome::Invalid(DiscardReason::Internal) => write!(f, "internal error"),
             Outcome::Invalid(reason) => write!(f, "invalid data ({reason})"),
             Outcome::Abuse => write!(f, "abuse limit reached"),