+++ tests/integration/test_metric_stats.py
@@ -173,13 +173,21 @@ def test_metric_stats_with_limit_surpassed(
     project_config["config"]["metrics"] = {
         "cardinalityLimits": [
             {
-                "id": "custom-limit",
+                "id": "org-limit",
+                "window": {"windowSeconds": 1, "granularitySeconds": 1},
+                "report": True,
+                "limit": 0,
+                "scope": "org",
+                "namespace": "custom",
+            },
+            {
+                "id": "name-limit",
                 "window": {"windowSeconds": 1, "granularitySeconds": 1},
                 "report": True,
                 "limit": 0,
                 "scope": "name",
                 "namespace": "custom",
-            }
+            },
         ]
     }
 
@@ -196,6 +204,7 @@ def test_metric_stats_with_limit_surpassed(
         "mri.type": "d",
         "mri.namespace": "custom",
         "outcome.id": "6",
+        "outcome.reason": "name-limit",
     }
     assert metrics.volume["d:custom/baz@none"]["org_id"] == 0
     assert metrics.volume["d:custom/baz@none"]["project_id"] == project_id
@@ -205,6 +214,7 @@ def test_metric_stats_with_limit_surpassed(
         "mri.type": "d",
         "mri.namespace": "custom",
         "outcome.id": "6",
+        "outcome.reason": "name-limit",
     }
     assert metrics.volume["s:custom/bar@none"]["org_id"] == 0
     assert metrics.volume["s:custom/bar@none"]["project_id"] == project_id
@@ -214,6 +224,7 @@ def test_metric_stats_with_limit_surpassed(
         "mri.type": "s",
         "mri.namespace": "custom",
         "outcome.id": "6",
+        "outcome.reason": "name-limit",
     }
     assert len(metrics.volume) == 3
     assert len(metrics.cardinality) == 0